The image does not match the html structure
The image shows numeric values, then square bullets and finally the snow image. However, the structure of your html is different, there is a <ul>, then a <ol> and then <ul> with class decorated. You can either change the image or you can change the ordering of the <ul> and <ol>.  I have changed the code to match the image by switching <ul> with <ol>.